requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aioschedule==0.5.2
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
loguru==0.7.0
magic-filter==1.0.9
multidict==6.0.4
numpy==1.25.0
pandas==2.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
sentry-sdk==1.26.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.3
yarl==1.9.2
